MODTLR-235 Unified `LoanActionService`
