+++ src/main/java/org/folio/service/LoanActionService.java
@@ -1,0 +1,11 @@
+package org.folio.service;
+
+import org.folio.domain.dto.ClaimItemReturnedRequest;
+import org.folio.domain.dto.DeclareClaimedReturnedItemAsMissingRequest;
+import org.folio.domain.dto.DeclareItemLostRequest;
+
+public interface LoanActionService {
+  void declareItemLost(DeclareItemLostRequest request);
+  void claimItemReturned(ClaimItemReturnedRequest request);
+  void declareClaimedReturnedItemMissing(DeclareClaimedReturnedItemAsMissingRequest request);
+}
