--- src/main/java/org/folio/service/impl/DeclareItemLostServiceImpl.java
@@ -1,46 +1,0 @@
-package org.folio.service.impl;
-
-import org.folio.client.feign.CirculationErrorForwardingClient;
-import org.folio.domain.dto.DeclareItemLostRequest;
-import org.folio.domain.dto.Loan;
-import org.folio.domain.mapper.CirculationMapper;
-import org.folio.repository.EcsTlrRepository;
-import org.folio.service.DeclareItemLostService;
-import org.folio.service.LoanService;
-import org.folio.service.RequestService;
-import org.folio.service.wrapper.LoanActionRequest;
-import org.folio.spring.service.SystemUserScopedExecutionService;
-import org.springframework.stereotype.Service;
-
-import lombok.extern.log4j.Log4j2;
-
-@Service
-@Log4j2
-public class DeclareItemLostServiceImpl extends AbstractLoanActionService<DeclareItemLostRequest>
-implements DeclareItemLostService {
-
-  public DeclareItemLostServiceImpl(
-    CirculationErrorForwardingClient circulationClient,
-    CirculationMapper circulationMapper,
-    LoanService loanService,
-    RequestService requestService,
-    EcsTlrRepository ecsTlrRepository,
-    SystemUserScopedExecutionService systemUserService) {
-
-    super(circulationClient, circulationMapper, loanService, requestService, ecsTlrRepository,
-      systemUserService);
-  }
-
-  public void declareItemLost(DeclareItemLostRequest request) {
-    process(LoanActionRequest.from(request));
-  }
-
-  @Override
-  protected void performActionInCirculation(Loan loan,
-    LoanActionRequest<DeclareItemLostRequest> actionRequest) {
-
-    log.info("declareItemLostInCirculation:: declaring item lost for loan {}", loan::getId);
-    circulationClient.declareItemLost(loan.getId(),
-      circulationMapper.toCirculationDeclareItemLostRequest(actionRequest.originalRequest()));
-  }
-}
